+++ packages/desktop/src/renderer/components/Settings/Settings.tsx
@@ -7,7 +7,7 @@ import { About } from '../widgets/Settings/About'
 import { Notifications } from './Tabs/Notifications/Notifications'
 import { Invite } from './Tabs/Invite/Invite'
 import { QRCode } from './Tabs/QRCode/QRCode'
-import { LeaveCommunity } from './Tabs/LeaveCommunity/LeaveCommunity'
+import { Theme } from './Tabs/Theme/Theme'
 
 import SettingsComponent from './SettingsComponent'
 
@@ -19,7 +19,7 @@ const Settings = () => {
     notifications: Notifications,
     invite: Invite,
     qrcode: QRCode,
-    leaveCommunity: LeaveCommunity,
+    theme: Theme,
   }
 
   const leaveCommunityModal = useModal(ModalName.leaveCommunity)