+++ packages/desktop/src/renderer/components/Settings/SettingsComponent.tsx
@@ -91,6 +91,14 @@ export const SettingsComponent: React.FC<SettingsComponentProps> = ({
               <ChevronRightIcon />
             </ListItemIcon>
           </ListItemButton>
+          <Divider />
+          <ListItemButton data-testid={'theme-settings-tab'} onClick={() => handleChange('theme')}>
+            <ListItemText>Theme</ListItemText>
+            <ListItemIcon>
+              <ChevronRightIcon />
+            </ListItemIcon>
+          </ListItemButton>
+          <Divider />
           {!isWindows && (
             <>
               <Divider />